move fixtures to conftest
