+++ python/cugraph-pyg/cugraph_pyg/data/__init__.py
@@ -11,7 +11,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from cugraph_pyg.utilities.api_tools import experimental_warning_wrapper
+from cugraph.utilities.api_tools import experimental_warning_wrapper
 
 from cugraph_pyg.data.cugraph_store import EXPERIMENTAL__CuGraphStore
