+++ python/cugraph-pyg/cugraph_pyg/tests/test_cugraph_store.py
@@ -25,89 +25,7 @@ import numpy as np
 
 import pytest
 
-from cugraph.gnn import FeatureStore
-
 from random import randint
-
-
-@pytest.fixture
-def basic_graph_1():
-    G = {
-        ("vt1", "pig", "vt1"): [
-            np.array([0, 0, 1, 2, 2, 3]),
-            np.array([1, 2, 4, 3, 4, 1]),
-        ]
-    }
-
-    N = {"vt1": 5}
-
-    F = FeatureStore()
-    F.add_data(np.array([100, 200, 300, 400, 500]), type_name="vt1", feat_name="prop1")
-
-    F.add_data(np.array([5, 4, 3, 2, 1]), type_name="vt1", feat_name="prop2")
-
-    return F, G, N
-
-
-@pytest.fixture
-def multi_edge_graph_1():
-    G = {
-        ("vt1", "pig", "vt1"): [np.array([0, 2, 3, 1]), np.array([1, 3, 1, 4])],
-        ("vt1", "dog", "vt1"): [np.array([0, 3, 4]), np.array([2, 2, 3])],
-        ("vt1", "cat", "vt1"): [
-            np.array([1, 2, 2]),
-            np.array([4, 3, 4]),
-        ],
-    }
-
-    N = {"vt1": 5}
-
-    F = FeatureStore()
-    F.add_data(np.array([100, 200, 300, 400, 500]), type_name="vt1", feat_name="prop1")
-
-    F.add_data(np.array([5, 4, 3, 2, 1]), type_name="vt1", feat_name="prop2")
-
-    return F, G, N
-
-
-@pytest.fixture
-def multi_edge_multi_vertex_graph_1():
-
-    G = {
-        ("brown", "horse", "brown"): [
-            np.array([0, 0]),
-            np.array([1, 2]),
-        ],
-        ("brown", "duck", "black"): [
-            np.array([1, 1, 2]),
-            np.array([1, 0, 1]),
-        ],
-        ("brown", "mongoose", "black"): [
-            np.array([2, 1]),
-            np.array([0, 1]),
-        ],
-        ("black", "cow", "brown"): [
-            np.array([0, 0]),
-            np.array([1, 2]),
-        ],
-        ("black", "snake", "black"): [
-            np.array([1]),
-            np.array([0]),
-        ],
-    }
-
-    N = {"brown": 3, "black": 2}
-
-    F = FeatureStore()
-    F.add_data(np.array([100, 200, 300]), type_name="brown", feat_name="prop1")
-
-    F.add_data(np.array([400, 500]), type_name="black", feat_name="prop1")
-
-    F.add_data(np.array([5, 4, 3]), type_name="brown", feat_name="prop2")
-
-    F.add_data(np.array([2, 1]), type_name="black", feat_name="prop2")
-
-    return F, G, N
 
 
 def test_tensor_attr():
